Reuse tracker service fake for service list test
